+++ PoporFoundation/Classes/NSObject/NSObject+pSwizzling.m
@@ -112,7 +112,7 @@
     NSString *replaceSelectorName = [NSString stringWithFormat:@"%@_%@",@"rep",originSelectorName];
     SEL repSelector = NSSelectorFromString(replaceSelectorName);
     for (NSInteger i=1; [destaClass respondsToSelector:repSelector]; i++) {
-        replaceSelectorName = [NSString stringWithFormat:@"%@%li_%@",@"rep",i,originSelectorName];
+        replaceSelectorName = [NSString stringWithFormat:@"%@%li_%@",@"rep", (long)i, originSelectorName];
         repSelector = NSSelectorFromString(replaceSelectorName);
     }
     IMP imp = imp_implementationWithBlock(callback);
